+++ packages/package-json-lint/src/rules/require-theme-context-versions.js
@@ -76,9 +76,9 @@ const versionSet = [
 ];
 
 module.exports = {
-  create: ({ ruleConfig, type, report }) => ({
+  create: ({ ruleConfig, projectType, report }) => ({
     dependencies: (dependencies) => requireVersionSet({
-      versionSet, dependencies, ruleConfig, type: type || 'module', report, lintId: 'require-theme-context-versions', messageString: 'theming context',
+      versionSet, dependencies, ruleConfig, projectType, report, lintId: 'require-theme-context-versions', messageString: 'theming context',
     }),
   }),
 };