+++ packages/package-json-lint/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 * Added
-  * Added `type` to allow users to flag project as `application`, `module` or `devModule`.
+  * Added `projectType` to allow users to flag project as `application`, `module` or `devModule`.
 
 ## 1.0.0 - (August 13, 2021)